requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==1.0.4
postgrest==0.10.8
bcrypt==4.0.1
PyJWT==2.8.0
openpyxl==3.1.2
pandas==2.1.1